requirements.txt
# --- Web & API ---
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7

# --- Dashboard ---
streamlit==1.31.0
plotly==5.19.0

# --- Database ---
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Data & Parsing ---
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0
openpyxl==3.1.2

# --- Security (Auth) ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# --- Utils ---
python-dotenv==1.0.1
loguru==0.7.2
apscheduler==3.10.4
pytest==8.0.0
